requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
pandas>=2.0.0
tqdm>=4.65.0
fastapi>=0.104.0
gunicorn>=21.2.0
uvicorn[standard]>=0.24.0
rank-bm25>=0.2.2
pymorphy2
nltk